+++ src/main/kotlin/models/Donation.kt
@@ -1,7 +1,7 @@
 package models
 
 data class Donation(val item: ItemDetails, val quantity: Int) : Comparable<Donation> {
-    val totalPrice: Int
+    val totalPrice: Long
         get() = item.price * quantity
 
     override fun compareTo(other: Donation): Int {
